Invert logic for _get_possible_moves
It is not intuitive to collate all of the possible moves and then select
what to play. When playing naturally there are some moves that one makes
without thinking. Therefore the most natural way to write the code is to
look for moves and play them in the same function rather than collecting
all of the possible moves first.

Also, this is a more useful paradigm for looking several moves ahead as
it requires less overhead.